I could not sleep so I decided to do some work on the project!
CategoryBean have been updated with the methods for getting comments, lists, toStrings and more.
methods for creation of test data is made as well in service.
Username is no always checked against it the list as lowercase so capital letters are no longer an issue in the username.

message.xhtml have been created and is used for showing the the currently selected message / news post.
most of the xhtml files have been updated with small <br/> tags and <h> tag changes.

UserBean have recieved an isAdmin method for comment deletion.
Category have gotten getters and setters for it's list of message and small fixes for initialization of the list.
CategoryBean have been updated so it correctly sets the messages list and comments list when a link is clicked.
Messages have gotten a title in addition to the text it already had. Corresponding getters, setters and toString methods have been updated to reflect this chance.
Service is mainly two sub private methods for creation of test data.

I think that's it AN EXTRA COMMIT WITH MESSAGE.XHTML WILL FOLLOW THIS ONE